python/requirements.txt
hubspot-api-client==3.8.2
python-dotenv==0.16.0
Flask==2.0.0
Werkzeug==2.2.2
